requirements.txt
flask
openai
gunicorn
whatsapp-api-client
python-dotenv
transformers
torch
